+++ extensions/ima/src/main/java/com/google/android/exoplayer2/ext/ima/ImaAdsLoader.java
@@ -82,6 +82,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Locale;
 import java.util.Set;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 
@@ -1912,7 +1913,9 @@ public final class ImaAdsLoader
   private static final class DefaultImaFactory implements ImaUtil.ImaFactory {
     @Override
     public ImaSdkSettings createImaSdkSettings() {
-      return ImaSdkFactory.getInstance().createImaSdkSettings();
+      ImaSdkSettings settings = ImaSdkFactory.getInstance().createImaSdkSettings();
+      settings.setLanguage(getImaLanguageCodeForDefaultLocale());
+      return settings;
     }
 
     @Override
@@ -1952,6 +1955,17 @@ public final class ImaAdsLoader
         Context context, ImaSdkSettings imaSdkSettings, AdDisplayContainer adDisplayContainer) {
       return ImaSdkFactory.getInstance()
           .createAdsLoader(context, imaSdkSettings, adDisplayContainer);
+    }
+
+    /**
+     * Returns a language code that's suitable for passing to {@link ImaSdkSettings#setLanguage} and
+     * corresponds to the device's {@link Locale#getDefault() default Locale}.
+     */
+    // TODO: It may be possible to define a better mapping onto IMA's supported language codes. See:
+    // https://developers.google.com/interactive-media-ads/docs/sdks/android/client-side/localization.
+    // IMA will fall back to its default language code ("en") if the value returned is unsupported.
+    private static String getImaLanguageCodeForDefaultLocale() {
+      return Util.splitAtFirst(Util.getSystemLanguageCodes()[0], "-")[0];
     }
   }
 }